% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{brewer_recycled}
\alias{brewer_recycled}
\title{Function to create a new color palette of size n starting from an existing RColorBrewer palette.}
\usage{
brewer_recycled(palette = "Dark2", n)
}
\arguments{
\item{palette}{name of a RColorBrewer palette (e.g., "Dark2")}

\item{n}{numeric integer indicating the number of colors in the new palette}
}
\value{
a new color palette of size equal to n
}
\description{
Function to create a new color palette of size n starting from an existing RColorBrewer palette.
}
