% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{c_distribution}
\alias{c_distribution}
\title{Generate a lineage and subtype distance distribution plot}
\usage{
c_distribution(dist_top_n, ann)
}
\arguments{
\item{dist_top_n}{neighbors dataframe}

\item{ann}{annotation file of tumors and cell lines}
}
\value{
lineage and subtype distance distribution plot
}
\description{
Generate a lineage and subtype distance distribution plot
}
