% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celligner_global_param.R
\docType{data}
\name{celligner_global}
\alias{celligner_global}
\title{Global parameters for expression data}
\format{
An object of class \code{list} of length 13.
}
\usage{
celligner_global
}
\description{
Take from https://github.com/broadinstitute/celligner/blob/d9c9246f8a1b6885d07f2f28bbdca24253e57cf1/R/global_params.R
}
\keyword{datasets}
