% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celligner_helpers.R
\name{check_NAs}
\alias{check_NAs}
\title{Helpers Function of Celligner}
\usage{
check_NAs(mat)
}
\arguments{
\item{mat:}{matrix of gene expression data that is samples by genes}
}
\value{
matrix of gene expression data, removing samples that have NAs
}
\description{
Take from https://github.com/broadinstitute/celligner/blob/d9c9246f8a1b6885d07f2f28bbdca24253e57cf1/R/Celligner_helpers.R
}
\details{
check for NAs in the expression data and remove samples with NAs
}
