% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celligner_method.R,
%   R/Create_Seurat_Object_upload.R
\name{create_Seurat_object}
\alias{create_Seurat_object}
\title{Method to create seurat objects given an expression matrix and annotation table}
\usage{
create_Seurat_object(mat, ann, type = NULL)

create_Seurat_object(mat, ann, type = NULL)
}
\arguments{
\item{mat}{matrix}

\item{ann}{annotation file: Expects column 'sampleID' which matches the rownames of mat.}

\item{type}{optional parameter: string specifying the data type of the current data (ex. 'tumor'), which is added to the annotation matrix.}

\item{exp_mat:}{matrix of samples by genes, where genes are ensembl gene IDs. Data should be log2(X+1) TPM data.}

\item{ann:}{matrix of sample anntoations. Expects column 'sampleID' which matches the rownames of exp_mat.}

\item{type:}{optional parameter, string specifying the data type of the current data (ex. 'tumor'), which is added to the annotation matrix.}
}
\value{
Seurat object with scaled expression data and annotations stored in meta.data

a 'v3' assay version Seurat object
}
\description{
create Seurat object of expression data and annotations and run dimensionality reduction.
Dimensionality reductions will be run with the parameters (n_PC_dims, umap_n_neighbors, umap_min_dist, distance_metric) specified in celligner_global.
}
