% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSeuObj_signature.R
\name{createSeuObj_signature}
\alias{createSeuObj_signature}
\title{Method to create seurat objects given Mutational signature matrix and annotation table}
\usage{
createSeuObj_signature(mat, ann, type = NULL)
}
\arguments{
\item{mat}{matrix of mutational signature data}

\item{ann}{annotation file: Expects column 'sampleID' which matches the rownames of mat.}

\item{type}{optional parameter: string specifying the data type of the current data (ex. 'tumor'), which is added to the annotation matrix.}
}
\value{
a Seurat object with mutational signature data
}
\description{
Adapted from: https://github.com/broadinstitute/celligner/blob/d9c9246f8a1b6885d07f2f28bbdca24253e57cf1/R/Celligner_methods.R
}
