% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_neighbors.R
\name{find_neighbors}
\alias{find_neighbors}
\title{Create a data frame with the neighbors and related annotation of a query sample/s}
\usage{
find_neighbors(
  combined_mat,
  reduced_mat,
  input_sample,
  k,
  ann,
  query_type,
  query_lineage = "All"
)
}
\arguments{
\item{combined_mat}{combined_mat matrix samples x genes of corrected data by MNN}

\item{reduced_mat}{dimensionally reduced matrix (tSNE and UMAP): sample x features}

\item{input_sample}{a vector of one or multiple TCGA and/or CCLE IDs.}

\item{k}{number of the nearest neighbors}

\item{ann}{annotation file of tumors and cell lines}

\item{query_type}{type of the k neighbors samples (Tumor or Cell Line)}

\item{query_lineage}{limit the neighbors searching to that lineage/s}
}
\value{
an interactive plot that highlighting the tumor k nearest neighbors
}
\description{
Create a data frame with the neighbors and related annotation of a query sample/s
}
