% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_neighbors_tumor.R
\name{find_neighbors_tumor}
\alias{find_neighbors_tumor}
\title{Generate an interactive plot that highlighting the tumor k nearest neighbors for the user query}
\usage{
find_neighbors_tumor(
  combined_mat,
  reduced_mat,
  input_sample,
  k,
  ann,
  BNindex,
  sample_order
)
}
\arguments{
\item{combined_mat}{combined_mat matrix samples x genes of corrected data by MNN}

\item{reduced_mat}{dimensionally reduced matrix (tSNE and UMAP): sample x features}

\item{input_sample}{samples of TCGA or CCLE choosed by the user}

\item{k}{number of the nearest neighbors}

\item{ann}{annotation file of tumors and cell lines}

\item{BNindex}{a BiocNeighborIndex object containing precomputed index information}

\item{sample_order}{index of the original combined matrix}
}
\value{
an interactive plot that highlighting the tumor k nearest neighbors found by queryKNN
}
\description{
Generate an interactive plot that highlighting the tumor k nearest neighbors for the user query
}
