% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_alignment_plot}
\alias{get_alignment_plot}
\title{Create an interactive plot that highlighting k nearest neighbors for the choosed query}
\usage{
get_alignment_plot(
  reduced_mat,
  ann,
  input_sample = NULL,
  dist_top_n = NULL,
  annot_value = "lineage"
)
}
\arguments{
\item{reduced_mat}{dimensionally reduced matrix (tSNE and UMAP): sample x features}

\item{ann}{annotation file of tumors and cell lines}

\item{input_sample}{a vector of one or multiple TCGA and/or CCLE IDs.}

\item{dist_top_n}{neighbors dataframe}

\item{annot_value}{indicates the annotation column from the annotation table 'ann'. Values acceptes are: 'lineage' or 'subtype'}
}
\value{
an interactive plot that highlighting the tumor k nearest neighbors
}
\description{
Create an interactive plot that highlighting k nearest neighbors for the choosed query
}
