% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_names_with_stripped.R
\name{get_all_name_strpp}
\alias{get_all_name_strpp}
\title{Get the stripped cell line name and the tumors ID of the user selected combined matrix}
\usage{
get_all_name_strpp(combined_mat, ann, sample_info)
}
\arguments{
\item{combined_mat}{matrix samples x genes of corrected data by MNN}

\item{ann}{annotation file of tumors and cell lines}
}
\value{
a vector with the stripped cell line name and the tumors ID
}
\description{
Get the stripped cell line name and the tumors ID of the user selected combined matrix
}
