% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_CL_strp_names}
\alias{get_CL_strp_names}
\title{Get the stripped cell line name of the user selected combined matrix}
\usage{
get_CL_strp_names(combined_mat, ann)
}
\arguments{
\item{combined_mat}{matrix samples x genes of corrected data by MNN:}

\item{ann}{annotation file of tumors and cell lines}
}
\value{
a vector with the stripped cell line name
}
\description{
Get the stripped cell line name of the user selected combined matrix
}
