% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_confusion_matrix.R
\name{get_confusion_matrix}
\alias{get_confusion_matrix}
\title{Function to create a confusion matrix (lineage TCGA x lineage CCLE)}
\usage{
get_confusion_matrix(combined_mat, ann_multiomics)
}
\arguments{
\item{combined_mat}{matrix with MNN corrected data: samples x genes}

\item{ann_multiomics}{annotation file of all the omics sample}
}
\value{
a list which contains prop_agree scores, confusion matrixs and a dataframe with the top 25 neighbors
}
\description{
Function to create a confusion matrix (lineage TCGA x lineage CCLE)
}
