% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_confusion_matrix_shiny.R
\name{get_confusion_matrix_shiny}
\alias{get_confusion_matrix_shiny}
\title{Get the stripped cell line name and the tumors ID of the user selected combined matrix}
\usage{
get_confusion_matrix_shiny(mat_m3)
}
\arguments{
\item{mat_m3}{dataframe of lineage frequencies}
}
\value{
lineages heatmap
}
\description{
Get the stripped cell line name and the tumors ID of the user selected combined matrix
}
