% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist_eu.R
\name{get_dist_eu}
\alias{get_dist_eu}
\title{Function to calculate the Euclidean distance for each TCGA sample form each CL sample}
\usage{
get_dist_eu(mnn_res, CCLE_cor)
}
\arguments{
\item{mnn_res}{TCGA matrix corrected by MNN}

\item{CCLE_cor}{CCLE matrix with the frist 4 cPCs regressed out}
}
\value{
the euclidean distance for each TCGA sample from each CL sample
}
\description{
Function to calculate the Euclidean distance for each TCGA sample form each CL sample
}
