% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist_eu_foreach_parallel.R
\name{get_fastdist_eu}
\alias{get_fastdist_eu}
\title{Function to calculate in parallel the Euclidean distance for each TCGA sample form each CL sample}
\usage{
get_fastdist_eu(mnn_res, CCLE_cor, cores)
}
\arguments{
\item{mnn_res}{TCGA matrix corrected by MNN}

\item{CCLE_cor}{CCLE matrix with the frist 4 cPCs regressed out}

\item{core}{number of core}
}
\value{
a list of a list: the euclidean distance for each TCGA sample from each CL sample
}
\description{
Function to calculate in parallel the Euclidean distance for each TCGA sample form each CL sample
}
