% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_heatmap_signature.R
\name{get_heatmap_signature}
\alias{get_heatmap_signature}
\title{Function to generate confusion matrix, Celligner_based_plot, proportion of agreement based on correlation and euclidean distance for mutational signature data}
\usage{
get_heatmap_signature(
  mnn_param,
  CCLE_cor,
  TCGA_cor,
  TCGA_ann,
  CCLE_ann,
  comb_ann,
  subset_genes
)
}
\arguments{
\item{mnn_param}{a vector of value for MNN param (k1,k2,ndist)}

\item{CCLE_cor}{CCLE matrix with the frist 4 cPCs regressed out}

\item{TCGA_cor}{TCGA matrix with the frist 4 cPCs regressed out}

\item{TCGA_ann}{TCGA annotation file}

\item{CCLE_ann}{CCLE annotation file}

\item{comb_ann}{TCGA e CCLE annotation file}

\item{subset_genes}{Vector of genes that have higher variance in both dataset}
}
\value{
a list that contains the confusion matrix, celligner_based plot, proportion of agreement based on correlation and euclidean distance
}
\description{
Function to generate confusion matrix, Celligner_based_plot, proportion of agreement based on correlation and euclidean distance for mutational signature data
}
