% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metasample_ann_both_both.R
\name{get_metasample_ann_both_both}
\alias{get_metasample_ann_both_both}
\title{Create a datatable of CL and tumor k nearest neighbors for the metasample sorted by distance, with samples information: lineage and subtype}
\usage{
get_metasample_ann_both_both(
  combined_mat,
  reduced_mat,
  selected_samples,
  n,
  ann
)
}
\arguments{
\item{combined_mat}{combined_mat matrix samples x genes of corrected data by MNN}

\item{reduced_mat}{dimensionally reduced matrix (tSNE and UMAP): sample x features}

\item{selected_samples}{vector of samples that will compone the metasample}

\item{n}{number of the nearest neighbors}

\item{ann}{annotation file of tumors and cell lines}
}
\value{
a datatable of CL and tumor samples with lineage and subtype information
}
\description{
Create a datatable of CL and tumor k nearest neighbors for the metasample sorted by distance, with samples information: lineage and subtype
}
