% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MOFA_fastdist_eu.R
\name{get_MOFA_fastdist_eu}
\alias{get_MOFA_fastdist_eu}
\title{Function to calculate in parallel the Euclidean distance for each TCGA sample from each CL sample for MOFA2 matrix}
\usage{
get_MOFA_fastdist_eu(MOFA_mat)
}
\arguments{
\item{MOFA_mat}{multiomics matrix of MOFA2 factor}
}
\value{
a list of a list: the euclidean distance for each TCGA sample from each CL sample for MOFA2 matrix
}
\description{
Function to calculate in parallel the Euclidean distance for each TCGA sample from each CL sample for MOFA2 matrix
}
