% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neighbors_table_tumor.R
\name{get_neighbors_table_tumor}
\alias{get_neighbors_table_tumor}
\title{Create a datatable of tumor k nearest neighbors from the user query sorted by distance, with samples information: lineage and subtype}
\usage{
get_neighbors_table_tumor(
  combined_mat,
  input_sample,
  k,
  ann,
  BNindex,
  sample_order
)
}
\arguments{
\item{combined_mat}{combined_mat matrix samples x genes of corrected data by MNN}

\item{input_sample}{samples of TCGA or CCLE choosed by the user}

\item{k}{number of the nearest neighbors}

\item{ann}{annotation file of tumors and cell lines}

\item{BNindex}{a BiocNeighborIndex object containing precomputed index information}

\item{sample_order}{index of the original combined matrix}
}
\value{
a datatable of tumor samples with lineage and subtype information
}
\description{
Create a datatable of tumor k nearest neighbors from the user query sorted by distance, with samples information: lineage and subtype
}
