% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_piechart}
\alias{get_piechart}
\title{Create a pie chart that illustrates the percentage of lineage or subtype belong to the k nearest neighbors}
\usage{
get_piechart(combined_mat, input_sample, type, ann, value, dist_top_n)
}
\arguments{
\item{combined_mat}{combined_mat matrix samples x genes of corrected data by MNN}

\item{input_sample}{a vector of one or multiple TCGA and/or CCLE IDs.}

\item{type}{type of the k neighbors samples (Tumor or Cell Line)}

\item{ann}{annotation file of tumors and cell lines}

\item{value}{string value to select to determine if the pie chart will be based on subtype percentage or lineage percentage}

\item{dist_top_n}{neighbors dataframe}
}
\value{
pie chart illustrating the distribution of k nearest neighbors subtype to the query
}
\description{
Create a pie chart that illustrates the percentage of lineage or subtype belong to the k nearest neighbors
}
