% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_piechart_both_2.R
\name{get_piechart_both_2}
\alias{get_piechart_both_2}
\title{Create a pie chart that illustrates the percentage of the lineages belong to CL and tumor k nearest neighbors samples for the metasample}
\usage{
get_piechart_both_2(combined_mat, selected_samples, n, ann)
}
\arguments{
\item{combined_mat}{combined_mat matrix samples x genes of corrected data by MNN}

\item{selected_samples}{vector of samples that will compone the metasample}

\item{n}{number of the nearest neighbors}

\item{ann}{annotation file of tumors and cell lines}

\item{reduced_mat}{dimensionally reduced matrix (tSNE and UMAP): sample x features}
}
\value{
a pie chart illustrating the distribution of the CL and TCGA k nearest neighbors lineages to the metasample
}
\description{
Create a pie chart that illustrates the percentage of the lineages belong to CL and tumor k nearest neighbors samples for the metasample
}
