% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_piechart_tumor.R
\name{get_piechart_tumor}
\alias{get_piechart_tumor}
\title{Create a pie chart that illustrates the percentage of the lineages belong to CL and tumor k nearest neighbors samples for the query sample}
\usage{
get_piechart_tumor(combined_mat, input_sample, k, ann, BNindex, sample_order)
}
\arguments{
\item{combined_mat}{combined_mat matrix samples x genes of corrected data by MNN}

\item{input_sample}{samples of TCGA or CCLE choosed by the user}

\item{k}{number of the nearest neighbors}

\item{ann}{annotation file of tumors and cell lines}

\item{BNindex}{a BiocNeighborIndex object containing precomputed index information}

\item{sample_order}{index of the original combined matrix}
}
\value{
a pie chart illustrating the distribution of the TCGA k nearest neighbors lineages to the query
}
\description{
Create a pie chart that illustrates the percentage of the lineages belong to CL and tumor k nearest neighbors samples for the query sample
}
