% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_agree_nofast_v6.R,
%   R/get_prop_agree_v6.R
\name{get_prop_agree_v6}
\alias{get_prop_agree_v6}
\title{Function to calculate the proportion of agreement based both on correlation and euclidean distance:
the euclidean distance will be not calculate in parallel}
\usage{
get_prop_agree_v6(
  mnn_param,
  CCLE_cor,
  TCGA_cor,
  TCGA_ann,
  CCLE_ann,
  comb_ann,
  subset_genes
)

get_prop_agree_v6(
  mnn_param,
  CCLE_cor,
  TCGA_cor,
  TCGA_ann,
  CCLE_ann,
  comb_ann,
  subset_genes
)
}
\arguments{
\item{mnn_param}{a vector of value for MNN param (k1,k2,ndist)}

\item{CCLE_cor}{CCLE matrix with the frist 4 cPCs regressed out}

\item{TCGA_cor}{TCGA matrix with the frist 4 cPCs regressed out}

\item{TCGA_ann}{TCGA annotation file}

\item{CCLE_ann}{CCLE annotation file}

\item{comb_ann}{TCGA e CCLE annotation file}

\item{subset_genes}{Vector of genes that have higher variance in both dataset}
}
\value{
a list that contains the value of prop_agree_dist, prop_agree_weighted_dist, prop_agree_cor and prop_agree_weigh_cor

a list that contains the value of prop_agree_dist, prop_agree_weighted_dist, prop_agree_cor and prop_agree_weigh_cor
}
\description{
Function to calculate the proportion of agreement based both on correlation and euclidean distance:
the euclidean distance will be not calculate in parallel

Function to calculate the proportion of agreement based both on correlation and euclidean distance
}
