% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function_2.R
\name{my_plotting}
\alias{my_plotting}
\title{Generate the interactive plot from the dimensionally reduced matrix}
\usage{
my_plotting(reduced_mat, ann)
}
\arguments{
\item{reduced_mat}{dimensionally reduced matrix (tSNE and UMAP): sample x features}

\item{ann}{annotation file of tumors and cell lines}
}
\value{
a interactive plot
}
\description{
Generate the interactive plot from the dimensionally reduced matrix
}
