% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function_4.R
\name{my_plotting_tSNE}
\alias{my_plotting_tSNE}
\title{Generate an interactive plot from the dimensionally reduced matrix}
\usage{
my_plotting_tSNE(reduced_mat, ann, omics_name)
}
\arguments{
\item{reduced_mat}{dimensionally reduced matrix (tSNE): sample x features}

\item{ann}{annotation file of tumors and cell lines}

\item{omics_name}{name of the shiny selected omics}
}
\value{
a interactive plot
}
\description{
Generate an interactive plot from the dimensionally reduced matrix
}
