% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_metasample_both_CL.R
\name{omics_metasample_both_CL}
\alias{omics_metasample_both_CL}
\title{Generate an interactive plot that highlighting the CL k nearest neighbors for the metasample}
\usage{
omics_metasample_both_CL(combined_mat, reduced_mat, selected_samples, n, ann)
}
\arguments{
\item{combined_mat}{combined_mat matrix samples x genes of corrected data by MNN}

\item{reduced_mat}{dimensionally reduced matrix (tSNE and UMAP): sample x features}

\item{selected_samples}{vector of samples that will compone the metasample}

\item{n}{number of the nearest neighbors}

\item{ann}{annotation file of tumors and cell lines}
}
\value{
an interactive plot that highlighting the CL k nearest neighbors for the metasample
}
\description{
Generate an interactive plot that highlighting the CL k nearest neighbors for the metasample
}
