% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_cluster_data.R
\name{signature_cluster_data}
\alias{signature_cluster_data}
\title{Method to take in a Seurat object and run default Seurat clustering algorithm for mutational signature data}
\usage{
signature_cluster_data(seu_obj)
}
\arguments{
\item{seu_obj}{Seurat object which contain mutational signature data}
}
\value{
a Seurat object
}
\description{
Adapted from: https://github.com/broadinstitute/celligner/blob/d9c9246f8a1b6885d07f2f28bbdca24253e57cf1/R/Celligner_methods.R
}
