% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_neighbors.R
\name{find_neighbors}
\alias{find_neighbors}
\title{Create an interactive plot that highlighting k nearest neighbors for the choosed query}
\usage{
find_neighbors(
  combined_mat,
  reduced_mat,
  input_sample = NULL,
  selected_samples = NULL,
  type,
  k,
  ann,
  omics_name,
  red_method,
  query_lineage
)
}
\arguments{
\item{combined_mat}{combined_mat matrix samples x genes of corrected data by MNN}

\item{reduced_mat}{dimensionally reduced matrix (tSNE and UMAP): sample x features}

\item{input_sample}{single TCGA or CCLE sample chosen by the user}

\item{selected_samples}{multiple TCGA or CCLE samples chosen by the user}

\item{type}{type of the k neighbors samples (Tumor or Cell Line)}

\item{k}{number of the nearest neighbors}

\item{ann}{annotation file of tumors and cell lines}

\item{omics_name}{name of the shiny selected omics}

\item{red_method}{dimensional reduction method}

\item{query_lineage}{limit the neighbors searching to that lineage/s}
}
\value{
an interactive plot that highlighting the tumor k nearest neighbors
}
\description{
Create an interactive plot that highlighting k nearest neighbors for the choosed query
}
