% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_piechart.R
\name{get_piechart}
\alias{get_piechart}
\title{Create a pie chart that illustrates the percentage of lineages belong to the k nearest neighbors}
\usage{
get_piechart(
  combined_mat,
  input_sample = NULL,
  selected_samples = NULL,
  type,
  k,
  ann,
  query_lineage
)
}
\arguments{
\item{combined_mat}{combined_mat matrix samples x genes of corrected data by MNN}

\item{input_sample}{single TCGA or CCLE sample chosen by the user}

\item{selected_samples}{multiple TCGA or CCLE samples chosen by the user}

\item{type}{type of the k neighbors samples (Tumor or Cell Line)}

\item{k}{number of the nearest neighbors}

\item{ann}{annotation file of tumors and cell lines}

\item{query_lineage}{limit the neighbors searching to that lineage/s}
}
\value{
a pie chart illustrating the distribution of k nearest neighbors lineages to the query
}
\description{
Create a pie chart that illustrates the percentage of lineages belong to the k nearest neighbors
}
